# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "ConversationDetail": {
        "additionalProperties": false,
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/ConversationView"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageView"
            },
            "title": "Messages",
            "type": "array"
          },
          "packet": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PacketView"
              },
              {
                "type": "null"
              }
            ]
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunView"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "conversation",
          "messages",
          "runs",
          "packet"
        ],
        "title": "ConversationDetail",
        "type": "object"
      },
      "ConversationView": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "organization_id",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "ConversationView",
        "type": "object"
      },
      "CreateConversationRequest": {
        "additionalProperties": false,
        "properties": {
          "title": {
            "default": "New planning conversation",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "title": "CreateConversationRequest",
        "type": "object"
      },
      "DoneEvent": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cursor": {
            "minimum": 0.0,
            "title": "Cursor",
            "type": "integer"
          },
          "event": {
            "const": "done",
            "default": "done",
            "title": "Event",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "cursor",
          "status"
        ],
        "title": "DoneEvent",
        "type": "object"
      },
      "ErrorEvent": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cursor": {
            "minimum": 0.0,
            "title": "Cursor",
            "type": "integer"
          },
          "event": {
            "const": "error",
            "default": "error",
            "title": "Event",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "cursor",
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorEvent",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MessageView": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "id",
          "conversation_id",
          "role",
          "content",
          "run_id",
          "created_at"
        ],
        "title": "MessageView",
        "type": "object"
      },
      "PacketUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "base_version": {
            "minimum": 0.0,
            "title": "Base Version",
            "type": "integer"
          },
          "content": {
            "additionalProperties": true,
            "title": "Content",
            "type": "object"
          },
          "idempotency_key": {
            "maxLength": 120,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          }
        },
        "required": [
          "base_version",
          "idempotency_key",
          "content"
        ],
        "title": "PacketUpdateRequest",
        "type": "object"
      },
      "PacketView": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "additionalProperties": true,
            "title": "Content",
            "type": "object"
          },
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "conversation_id",
          "version",
          "content",
          "created_at"
        ],
        "title": "PacketView",
        "type": "object"
      },
      "RunView": {
        "additionalProperties": false,
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "retry_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Of"
          },
          "runtime": {
            "enum": [
              "fixture",
              "provider"
            ],
            "title": "Runtime",
            "type": "string"
          },
          "status": {
            "enum": [
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "conversation_id",
          "status",
          "runtime",
          "retry_of",
          "created_at",
          "completed_at"
        ],
        "title": "RunView",
        "type": "object"
      },
      "SessionEvent": {
        "additionalProperties": false,
        "properties": {
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cursor": {
            "minimum": 0.0,
            "title": "Cursor",
            "type": "integer"
          },
          "event": {
            "const": "session",
            "default": "session",
            "title": "Event",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "runtime": {
            "default": "fixture",
            "enum": [
              "fixture",
              "provider"
            ],
            "title": "Runtime",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "cursor"
        ],
        "title": "SessionEvent",
        "type": "object"
      },
      "StatusEvent": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cursor": {
            "minimum": 0.0,
            "title": "Cursor",
            "type": "integer"
          },
          "event": {
            "const": "status",
            "default": "status",
            "title": "Event",
            "type": "string"
          },
          "message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Id"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "cursor",
          "status"
        ],
        "title": "StatusEvent",
        "type": "object"
      },
      "TextDeltaEvent": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cursor": {
            "minimum": 0.0,
            "title": "Cursor",
            "type": "integer"
          },
          "delta": {
            "title": "Delta",
            "type": "string"
          },
          "event": {
            "const": "text_delta",
            "default": "text_delta",
            "title": "Event",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "cursor",
          "delta"
        ],
        "title": "TextDeltaEvent",
        "type": "object"
      },
      "TurnStreamRequest": {
        "additionalProperties": false,
        "properties": {
          "after_cursor": {
            "default": -1,
            "minimum": -1.0,
            "title": "After Cursor",
            "type": "integer"
          },
          "client_message_id": {
            "maxLength": 120,
            "minLength": 8,
            "title": "Client Message Id",
            "type": "string"
          },
          "message": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "reconnect_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reconnect Run Id"
          },
          "retry_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Of"
          },
          "runtime": {
            "default": "fixture",
            "enum": [
              "fixture",
              "provider"
            ],
            "title": "Runtime",
            "type": "string"
          }
        },
        "required": [
          "message",
          "client_message_id"
        ],
        "title": "TurnStreamRequest",
        "type": "object"
      },
      "UsageEvent": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cursor": {
            "minimum": 0.0,
            "title": "Cursor",
            "type": "integer"
          },
          "event": {
            "const": "usage",
            "default": "usage",
            "title": "Event",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageSnapshot"
          }
        },
        "required": [
          "run_id",
          "cursor",
          "usage"
        ],
        "title": "UsageEvent",
        "type": "object"
      },
      "UsageSnapshot": {
        "additionalProperties": false,
        "properties": {
          "actual_service_tier": {
            "default": "fixture",
            "title": "Actual Service Tier",
            "type": "string"
          },
          "cache_write_tokens": {
            "minimum": 0.0,
            "title": "Cache Write Tokens",
            "type": "integer"
          },
          "cached_input_tokens": {
            "minimum": 0.0,
            "title": "Cached Input Tokens",
            "type": "integer"
          },
          "cost_microusd": {
            "default": 0,
            "minimum": 0.0,
            "title": "Cost Microusd",
            "type": "integer"
          },
          "evidence_class": {
            "enum": [
              "synthetic",
              "development"
            ],
            "title": "Evidence Class",
            "type": "string"
          },
          "input_tokens": {
            "minimum": 0.0,
            "title": "Input Tokens",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_tokens": {
            "minimum": 0.0,
            "title": "Output Tokens",
            "type": "integer"
          },
          "price_version": {
            "default": "synthetic-zero-v1",
            "title": "Price Version",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "reasoning_tokens": {
            "minimum": 0.0,
            "title": "Reasoning Tokens",
            "type": "integer"
          },
          "total_tokens": {
            "minimum": 0.0,
            "title": "Total Tokens",
            "type": "integer"
          }
        },
        "required": [
          "provider",
          "model",
          "input_tokens",
          "cached_input_tokens",
          "cache_write_tokens",
          "output_tokens",
          "reasoning_tokens",
          "total_tokens",
          "evidence_class"
        ],
        "title": "UsageSnapshot",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Clio's bounded planning-agent foundation.",
    "title": "Clio API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/conversations": {
      "get": {
        "operationId": "list_conversations_api_v1_conversations_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-clio-organization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clio-Organization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConversationView"
                  },
                  "title": "Response List Conversations Api V1 Conversations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Conversations",
        "tags": [
          "application-shell"
        ]
      },
      "post": {
        "operationId": "create_conversation_api_v1_conversations_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-clio-organization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clio-Organization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Conversation",
        "tags": [
          "application-shell"
        ]
      }
    },
    "/api/v1/conversations/{conversation_id}": {
      "get": {
        "operationId": "get_conversation_api_v1_conversations__conversation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-clio-organization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clio-Organization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Conversation",
        "tags": [
          "application-shell"
        ]
      }
    },
    "/api/v1/conversations/{conversation_id}/packet": {
      "put": {
        "operationId": "update_packet_api_v1_conversations__conversation_id__packet_put",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-clio-organization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clio-Organization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PacketUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PacketView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Packet",
        "tags": [
          "application-shell"
        ]
      }
    },
    "/api/v1/conversations/{conversation_id}/turns/stream": {
      "post": {
        "operationId": "stream_turn_api_v1_conversations__conversation_id__turns_stream_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-clio-organization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clio-Organization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnStreamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "done": "#/components/schemas/DoneEvent",
                      "error": "#/components/schemas/ErrorEvent",
                      "session": "#/components/schemas/SessionEvent",
                      "status": "#/components/schemas/StatusEvent",
                      "text_delta": "#/components/schemas/TextDeltaEvent",
                      "usage": "#/components/schemas/UsageEvent"
                    },
                    "propertyName": "event"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SessionEvent"
                    },
                    {
                      "$ref": "#/components/schemas/StatusEvent"
                    },
                    {
                      "$ref": "#/components/schemas/TextDeltaEvent"
                    },
                    {
                      "$ref": "#/components/schemas/UsageEvent"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorEvent"
                    },
                    {
                      "$ref": "#/components/schemas/DoneEvent"
                    }
                  ],
                  "title": "Response 200 Stream Turn Api V1 Conversations  Conversation Id  Turns Stream Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Turn",
        "tags": [
          "application-shell"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_run_api_v1_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-clio-organization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Clio-Organization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Run Api V1 Runs  Run Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run",
        "tags": [
          "application-shell"
        ]
      }
    },
...
  }
}